Adapt layout of AMQP update to VILLAScontroller protocol, add uptime in IC creation and updates via AMQP; need to check data type of uptime #31

+++ database/models.go
@@ -129,7 +129,7 @@ type InfrastructureComponent struct {
 	// Type of IC (RTDS, VILLASnode, RTDS, etc.)
 	Type string `json:"type" gorm:"default:''"`
 	// Uptime of the IC
-	Uptime int `json:"uptime" gorm:"default:0"`
+	Uptime float64 `json:"uptime" gorm:"default:0"`
 	// State of the IC
 	State string `json:"state" gorm:"default:''"`
 	// Time of last state update
